requirements-extractors.txt
pypdf>=5.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
xlrd>=2.0.1
pywin32>=306; platform_system == "Windows"
textract>=1.6.5; platform_system != "Windows"
